Add some linter tooling

+++ contracts/IRevoBounty.sol
@@ -4,11 +4,19 @@ import "./openzeppelin-solidity/contracts/IERC20.sol";
 
 struct TokenAmount {
     IERC20 token;
-    uint amount;
+    uint256 amount;
 }
 
 interface IRevoBounty {
-    function calculateBountyFee(TokenAmount[] calldata interestAccrued) external view returns (TokenAmount[] memory);
-    function calculateReserveFee(TokenAmount[] calldata interestAccrued) external view returns (TokenAmount[] memory);
+    function calculateBountyFee(TokenAmount[] calldata interestAccrued)
+        external
+        view
+        returns (TokenAmount[] memory);
+
+    function calculateReserveFee(TokenAmount[] calldata interestAccrued)
+        external
+        view
+        returns (TokenAmount[] memory);
+
     function issueAdditionalBounty(address recipient) external;
 }
